requirements.txt
qdrant-client>=1.7.1
pandas>=2.1.4
openpyxl>=3.1.2
scikit-learn>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0